+++ LinearAlgebraAssessment/space_rock.h
@@ -3,14 +3,14 @@
 
 class space_rock : public base_object
 {
-	matrix<float> matrix_{};
+	matrix<double> matrix_{};
 public:
-	space_rock(const std::vector<vector_3d<float>>& vector_draw_points)
+	space_rock(const std::vector<vector_3d<double>>& vector_draw_points)
 	{
 		matrix_.init_with_position_vectors(vector_draw_points);
 	}
 
-	matrix<float>& get_matrix() override { return matrix_; }
+	matrix<double>& get_matrix() override { return matrix_; }
 	void update(delta_time dt) override{}
 	void handle_event(SDL_Event& e) override{}
 };
